moved computationals to a computation package and TreeVisitors to a TreeVisitor package
