+++ src/main/java/de/featjar/feature/model/analysis/computation/ComputeFeatureAverageNumberOfChildren.java
@@ -18,34 +18,33 @@
  *
  * See <https://github.com/FeatureIDE/FeatJAR-feature-model> for further information.
  */
-package de.featjar.feature.model.analysis;
+package de.featjar.feature.model.analysis.computation;
 
 import de.featjar.base.computation.AComputation;
 import de.featjar.base.computation.Dependency;
 import de.featjar.base.computation.IComputation;
 import de.featjar.base.computation.Progress;
 import de.featjar.base.data.Result;
-import de.featjar.formula.assignment.BooleanAssignmentList;
+import de.featjar.base.tree.Trees;
+import de.featjar.feature.model.IFeatureTree;
+import de.featjar.feature.model.analysis.visitor.TreeAvgChildrenCounterTreeVisitor;
 import java.util.List;
 
 /**
- * Compute how many assignments are in the assignmentList
+ * Calculates the number of features that have no child features.
  *
- * @author Mohammad Khair Almekkawi
- * @author Florian Beese
+ * @author Benjamin von Holt
  */
-public class ComputeNumberConfigurations extends AComputation<Integer> {
+public class ComputeFeatureAverageNumberOfChildren extends AComputation<Double> {
+    protected static final Dependency<IFeatureTree> FEATURE_TREE = Dependency.newDependency(IFeatureTree.class);
 
-    public static final Dependency<BooleanAssignmentList> BOOLEAN_ASSIGNMENT_LIST =
-            Dependency.newDependency(BooleanAssignmentList.class);
-
-    public ComputeNumberConfigurations(IComputation<BooleanAssignmentList> booleanAssignmentList) {
-        super(booleanAssignmentList);
+    public ComputeFeatureAverageNumberOfChildren(IComputation<IFeatureTree> featureTree) {
+        super(featureTree);
     }
 
     @Override
-    public Result<Integer> compute(List<Object> dependencyList, Progress progress) {
-        BooleanAssignmentList booleanAssigmenAssignmentList = BOOLEAN_ASSIGNMENT_LIST.get(dependencyList);
-        return Result.of(booleanAssigmenAssignmentList.getAll().size());
+    public Result<Double> compute(List<Object> dependencyList, Progress progress) {
+        IFeatureTree tree = FEATURE_TREE.get(dependencyList);
+        return Trees.traverse(tree, new TreeAvgChildrenCounterTreeVisitor());
     }
 }
