+++ src/main/java/de/featjar/feature/model/cli/PrintSampleStatistics.java
@@ -29,11 +29,11 @@ import de.featjar.base.data.Result;
 import de.featjar.base.io.IO;
 import de.featjar.feature.model.IFeatureModel;
 import de.featjar.feature.model.analysis.AnalysisTree;
-import de.featjar.feature.model.analysis.ComputeDistributionFeatureSelections;
-import de.featjar.feature.model.analysis.ComputeFeatureCounter;
-import de.featjar.feature.model.analysis.ComputeNumberConfigurations;
-import de.featjar.feature.model.analysis.ComputeNumberVariables;
-import de.featjar.feature.model.analysis.ComputeUniformity;
+import de.featjar.feature.model.analysis.computation.ComputeDistributionFeatureSelections;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureCounter;
+import de.featjar.feature.model.analysis.computation.ComputeNumberConfigurations;
+import de.featjar.feature.model.analysis.computation.ComputeNumberVariables;
+import de.featjar.feature.model.analysis.computation.ComputeUniformity;
 import de.featjar.feature.model.io.FeatureModelFormats;
 import de.featjar.feature.model.io.csv.CSVAnalysisFormat;
 import de.featjar.feature.model.io.json.JSONAnalysisFormat;
