+++ src/test/java/de/featjar/feature/model/analysis/SamplePropertiesTest.java
@@ -28,6 +28,11 @@ import de.featjar.base.computation.IComputation;
 import de.featjar.feature.model.FeatureModel;
 import de.featjar.feature.model.IFeatureModel;
 import de.featjar.feature.model.TestDataProvider;
+import de.featjar.feature.model.analysis.computation.ComputeDistributionFeatureSelections;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureCounter;
+import de.featjar.feature.model.analysis.computation.ComputeNumberConfigurations;
+import de.featjar.feature.model.analysis.computation.ComputeNumberVariables;
+import de.featjar.feature.model.analysis.computation.ComputeUniformity;
 import de.featjar.feature.model.transformer.ComputeFormula;
 import de.featjar.formula.VariableMap;
 import de.featjar.formula.assignment.BooleanAssignment;
